+++ src/dms_provider_bridge/drivers/alfresco.py
@@ -18,12 +18,12 @@ from dms_provider_bridge.models.bridge import BridgeAuthContext
 from dms_provider_bridge.models.item import DmsItem
 from dms_provider_bridge.models.listing import ListingResult
 from dms_provider_bridge.models.operation import OperationResult
-from dms_provider_bridge.drivers.base import Provider
+from dms_provider_bridge.drivers.tc_vfs_contract import TcVfsContract
 from dms_provider_bridge.drivers import alfresco_config, alfresco_items, alfresco_share, alfresco_versioning
 from dms_provider_bridge.services.auth_resolver import resolve_effective_auth
 
 
-class AlfrescoProvider(Provider):
+class AlfrescoProvider(TcVfsContract):
     name = "alfresco"
     upstream_auth_scheme = "ticket"