+++ src/dms_provider_bridge/drivers/webdav.py
@@ -16,14 +16,14 @@ from dms_provider_bridge.models.bridge import BridgeAuthContext
 from dms_provider_bridge.models.item import DmsItem
 from dms_provider_bridge.models.listing import ListingResult
 from dms_provider_bridge.models.operation import OperationResult
-from dms_provider_bridge.drivers.base import Provider
+from dms_provider_bridge.drivers.tc_vfs_contract import TcVfsContract
 from dms_provider_bridge.services.auth_resolver import resolve_effective_auth
 
 
 DAV_NS = "{DAV:}"
 
 
-class WebdavProvider(Provider):
+class WebdavProvider(TcVfsContract):
     name = "webdav"
     upstream_auth_scheme = "basic"